requirements.txt
PySide6>=6.5.0
yt-dlp>=2023.11.16
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0  # 用于生成测试覆盖率报告
pytest-benchmark>=4.0.0  # 用于性能基准测试
pytest-xdist>=3.5.0  # 用于并行测试
pyyaml>=6.0.0
requests>=2.31.0
python-socks>=2.4.3  # 用于支持SOCKS代理
urllib3>=2.0.7
certifi>=2023.7.22   # 用于SSL证书验证
beautifulsoup4>=4.12.0
lxml>=4.9.0
typing-extensions>=4.5.0 